Refactor constraint view models

+++ base/plugins/scenario/source/Process/Temporal/TemporalScenarioProcessPresenter.hpp
@@ -103,7 +103,7 @@ class TemporalScenarioProcessPresenter : public ProcessPresenterInterface
 		id_type<EventModel> m_currentlySelectedEvent{};
 		int m_pointedEvent{0};
 
-		long m_millisecPerPixel{1};
+		double m_millisecPerPixel{1};
 
 		int m_horizontalZoomSliderVal{};
 };
